separate out the vectorise over dicts
